% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_ROC.R
\name{auroc}
\alias{auroc}
\alias{auroc.mixo_plsda}
\alias{auroc.mixo_splsda}
\alias{auroc.mint.plsda}
\alias{auroc.mint.splsda}
\alias{auroc.sgccda}
\title{Area Under the Curve (AUC) and Receiver Operating Characteristic (ROC)
curves for supervised classification}
\usage{
auroc(object, ...)
}
\arguments{
\item{object}{Object of class inherited from one of the following supervised
analysis function: "plsda", "splsda", "mint.plsda", "mint.splsda",
"block.splsda" or "wrapper.sgccda"}

\item{...}{external optional arguments for plotting}

\item{newdata}{numeric matrix of predictors, by default set to the training
data set (see details).}

\item{outcome.test}{Either a factor or a class vector for the discrete
outcome, by default set to the outcome vector from the training set (see
details).}

\item{study.test}{For MINT objects, grouping factor indicating which samples
of \code{newdata} are from the same study. Overlap with \code{object$study}
are allowed.}

\item{multilevel}{Sample information when a newdata matrix is input and when
multilevel decomposition for repeated measurements is required. A numeric
matrix or data frame indicating the repeated measures on each individual,
i.e. the individuals ID. See examples in \code{splsda}.}

\item{plot}{Whether the ROC curves should be plotted, by default set to TRUE
(see details).}

\item{roc.comp}{Specify the component (integer) for which the ROC will be
plotted from the multivariate model, default to 1.}

\item{roc.block}{Specify the block number (integer) or the name of the block
(set of characters) for which the ROC will be plotted for a block.plsda or
block.splsda object, default to 1.}

\item{roc.study}{Specify the study for which the ROC will be plotted for a
mint.plsda or mint.splsda object, default to "global".}
}
\value{
Depending on the type of object used, a list that contains: The AUC
and Wilcoxon test pvalue for each 'one vs other' classes comparison
performed, either per component (splsda, plsda, mint.plsda, mint.splsda), or
per block and per component (wrapper.sgccda, block.plsda, blocksplsda).
}
\description{
Calculates the AUC and plots ROC for supervised objects from s/plsda,
mint.s/plsda and block.plsda, block.splsda or wrapper.sgccda.
}
\details{
For more than two classes in the categorical outcome Y, the AUC is
calculated as one class vs. the other and the ROC curves one class vs. the
others are output.

The ROC and AUC are calculated based on the predicted scores obtained from
the \code{predict} function applied to the multivariate methods
(\code{predict(object)$predict}). Our multivariate supervised methods
already use a prediction threshold based on distances (see \code{predict})
that optimally determine class membership of the samples tested. As such AUC
and ROC are not needed to estimate the performance of the model (see
\code{perf}, \code{tune} that report classification error rates). We provide
those outputs as complementary performance measures.

The pvalue is from a Wilcoxon test between the predicted scores between one
class vs the others.

External independent data set (\code{newdata}) and outcome
(\code{outcome.test}) can be input to calculate AUROC. The external data set
must have the same variables as the training data set (\code{object$X}).

If \code{newdata} is not provided, AUROC is calculated from the training
data set, and may result in overfitting (too optimistic results).

Note that for mint.plsda and mint.splsda objects, if \code{roc.study} is
different from "global", then \code{newdata}), \code{outcome.test} and
\code{sstudy.test} are not used.
}
\seealso{
\code{\link{tune}}, \code{\link{perf}}, and http://www.mixOmics.org
for more details.
}
\author{
Benoit Gautier, Francois Bartolo, Florian Rohart
}
\keyword{multivariate}
\keyword{regression}
