% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pca.R
\name{plot.pca}
\alias{plot.pca}
\title{plot methods for mixOmics}
\usage{
\method{plot}{pca}(x, ncomp = min(10, length(x$sdev)),
  type = "barplot", explained.var = TRUE, ...)
}
\arguments{
\item{x}{a \code{pca} object obtained from \code{pca} function.}

\item{ncomp}{number of PCs to show.}

\item{type}{type of the plot, either "barplot" or argument passed to \code{type} in base \code{plot}.}

\item{explained.var}{logical. Whether to show proportion of variance explained (TRUE) or the total variance (FALSE).}

\item{...}{other arguments passed to \code{plot}.}
}
\value{
A scree plot of explained variance by the study PCs.
}
\description{
Plot the explained variances from a pca object
}
\details{
Creates a scree plot of explained variance by the study PCs.
}
\examples{
#' \dontrun{
## successful: TRUE
## final: TRUE

library(mixOmics.data)
pca.res <- pca(multidrug$ABC.trans, ncomp = 10, scale = TRUE)
plot(pca.res)
#' }
}
\author{
Florian Rohart, Kim-Anh Lê Cao, Ignacio González, Al J Abadi
\code{\link{plotIndiv}}, \code{\link{pca}} and http://www.mixOmics.org
for more details.
}
\keyword{plot}
