% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipca.R
\name{sipca}
\alias{sipca}
\alias{sipca.MultiAssayExperiment}
\alias{sipca.SingleCellExperiment}
\alias{sipca.SummarizedExperiment}
\title{Independent Principal Component Analysis}
\usage{
{sipca}{default}(X, assay=NULL, ncomp = 3, mode = c("deflation","parallel"),
fun = c("logcosh", "exp"), scale = FALSE, max.iter = 200, tol = 1e-04,
keepX = rep(50,ncomp), w.init = NULL)

\method{sipca}{MultiAssayExperiment}(X, assay, ...)

\method{sipca}{SingleCellExperiment}(X, assay = "logcounts", ...)

\method{sipca}{SummarizedExperiment}(X, assay, ...)
}
\arguments{
\item{X}{a numeric matrix (or data frame) which provides the data for the
principal components analysis. It can contain missing values. Alternatively, a \code{MultiAssay-/Summarized-/SingleCellExperiment} object.}

\item{assay}{name or index of an assay from \code{X}.}

\item{...}{Not used.}

\item{keepX}{numeric vector of length ncomp, the number of variables to keep}
}
\value{
\code{pca} returns a list with class \code{"ipca"} containing the
following components: \item{ncomp}{the number of principal components used.}
\item{unmixing}{the unmixing matrix of size (ncomp x ncomp)}
\item{mixing}{the mixing matrix of size (ncomp x ncomp} \item{X}{the
centered data matrix} \item{x}{the principal components (with sparse
independent loadings)} \item{loadings}{the sparse independent loading
vectors} \item{kurtosis}{the kurtosis measure of the independent loading
vectors}
}
\description{
Performs sparse independent principal component analysis on the given data
matrix to enable variable selection.
}
\details{
See Details of ipca.

Soft thresholding is implemented on the independent loading vectors to
obtain sparse loading vectors and enable variable selection.
}
\examples{
#' \dontrun{
## successful: FALSE

library(mixOmics.data)

# implement IPCA on a microarray dataset
sipca.res <- sipca(liver.toxicity$gene, ncomp = 3, mode="deflation", keepX=c(50,50,50))
sipca.res
# samples representation
plotIndiv(sipca.res, ind.names = liver.toxicity$treatment[, 4],
          group = as.numeric(as.factor(liver.toxicity$treatment[, 4])))
plotIndiv(sipca.res, cex = 1,
          col = as.numeric(as.factor(liver.toxicity$treatment[, 4])),style="3d")

# variables representation
plotVar(sipca.res, cex = 2.5)
plotVar(sipca.res, rad.in = 0.5, cex = 1,style="3d", cutoff = 0.75)
#' }
}
\references{
Yao, F., Coquery, J. and Lê Cao, K.-A. (2011) Principal
component analysis with independent loadings: a combination of PCA and ICA.
(in preparation)

A. Hyvarinen and E. Oja (2000) Independent Component Analysis: Algorithms
and Applications, \emph{Neural Networks}, \bold{13(4-5)}:411-430

J L Marchini, C Heaton and B D Ripley (2010). fastICA: FastICA Algorithms to
perform ICA and Projection Pursuit. R package version 1.1-13.
}
\seealso{
\code{\link{ipca}}, \code{\link{pca}}, \code{\link{plotIndiv}},
\code{\link{plotVar}} and http://www.mixOmics.org for more details.
}
\author{
Fangzhou Yao and Jeff Coquery.
}
\keyword{algebra}
